Workaround for DevOps coercion
The `eq` and `ne` functions coerce their second argument to be of the
same type as the first:
https://docs.microsoft.com/en-us/azure/devops/pipelines/process/expressions#eq.
If the first argument is a parameter, and it is set to `false` or `true`
(as opposed to `'false'` or `'true'`), then the second argument which
we're comparing against will be coerced to a boolean. However, since any
non-empty string is coerced to true if you had this expression:

```
${{ if eq(parameters.foo, 'false') }}:
```

Then passing `foo` as `true` would result in the condition being
considered _true_ (since `'false'` => `true == true`), and passing `foo`
as `false` would result in the condition being considered _false_ by the
same reasoning. The exact opposite of what you wanted. See also
https://docs.microsoft.com/en-us/azure/devops/pipelines/process/expressions#type-casting

This PR flips the operand order for every invocation of `eq` and `ne`
such that the arguments are always considered as _strings_. This fixes
the behavior observed above, since `false` is cast to `'false'` and
`true` to `'true`'.

It's stupid.